### 2024-11-07 #7422 @kthyng
@weber1158
* can you provide installation instructions?
* What would you consider to be tests for users (and you) to verify the software is working correctly?
* Please reduce to length of your paper to up to about 1000 words (currently 1194). Paper format information here: https://joss.readthedocs.io/en/latest/paper.html. You can check paper length (and a few other items) with the command `@editorialbot check repository` though note it needs to be the first item in a comment to run correctly.
