### 2024-11-09 1cc4f48
Shortened paper.md, added installation instructions, and replaced old test script
