+++ Paper/paper.bib
@@ -98,45 +98,3 @@
   publisher={Kluwer Academic Publishers},
   doi = {10.1007/978-1-4020-2841-0}
 }
-
-@article{Whitney2010,
-  title={Abbreviations for names of rock-forming minerals},
-  author={Whitney, Donna L. and Evans, Bernard W.},
-  journal={American mineralogist},
-  volume={95},
-  number={1},
-  pages={185--187},
-  year={2010},
-  publisher={Mineralogical Society of America},
-  doi = {10.2138/am.2010.3371}
-}
-
-@article{daSilva2021,
-  title={Qmin--{A} machine learning-based application for processing and analysis of mineral chemistry data},
-  author={da Silva, Guilherme Ferreira and Ferreira, Marcos Vinicius and Costa, Iago Sousa Lima and Bernardes, Renato Borges and Mota, Carlos Eduardo Miranda and Jim{\'e}nez, Federico Alberto Cuadros},
-  journal={Computers \& Geosciences},
-  volume={157},
-  pages={104949},
-  year={2021},
-  publisher={Elsevier},
-  doi = {10.1016/j.cageo.2021.104949}
-}
-
-@article{Chawla2002,
-  title={SMOTE: synthetic minority over-sampling technique},
-  author={Chawla, Nitesh V and Bowyer, Kevin W and Hall, Lawrence O and Kegelmeyer, W Philip},
-  journal={Journal of Artificial Intelligence Research},
-  volume={16},
-  pages={321--357},
-  year={2002},
-  doi = {10.1613/jair.953}
-}
-
-@misc{Larsen2024,
-  author = {Larsen, Bjarke Skogstad},
-  title = {Synthetic {Minority} {Over-sampling} {Technique} ({SMOTE})},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://github.com/dkbsl/matlab_smote}}
-}
