+++ README.md
@@ -10,6 +10,25 @@ This repository includes several functions designed to quickly identify common m
 
 Moreover, this repository includes MATLAB functions for importing EDS x-ray spectral data (<small>`read_msa()`</small>) and visualizing the data (<small>`xray_plot()`</small>). Users may also import the metadata from SEM micrographs with the <small>`get_sem_metadata()`</small> function, and more.
 
+## Installation
+You can download the repository from the MATLAB Central File Exchange [![View my project on File Exchange](https://www.mathworks.com/matlabcentral/images/matlab-file-exchange.svg)](https://www.mathworks.com/matlabcentral/fileexchange/170771), or open it directly in your browser [![Open in MATLAB Online](https://www.mathworks.com/images/responsive/global/open-in-matlab-online.svg)](https://matlab.mathworks.com/open/fileexchange/v1?id=170771&file=Tests/matlab_test.mlx) (recommended). 
+
+The repository was developed in MATLAB Online, which uses the most up-to-date version of MATLAB. To ensure backwards compatability, it is recommended that users also utilize the functions in MATLAB Online.
+
+To add the EDS Classification functions to the default search path:
+
+1. Un-zip the downloaded folder. 
+
+2. Execute the following command in the MATLAB Command Window:
+
+```matlab
+pathtool
+```
+
+3. A popup menu should open. Click `Add Folder with Subfolders` and select the un-zipped main repository folder. 
+
+4. Finalize your choice by clicking `Save` or `Apply`.
+
 ## How to cite
 Weber, Austin M. (2024) EDS Classification Version #.#.# [Software]. GitHub. https://github.com/weber1158/eds-classification
 
