+++ Paper/supplement.md
@@ -8,7 +8,7 @@ by Austin M. Weber<sup>1,2</sup>
 ## The "Weber Algorithm": A novel machine learning classifier for categorizing minerals from EDS net intensity data
 ### Reference Minerals
 
-A total of 20 mineral standards were prepared in order to collect EDS data for training a novel mineral classification model using supervised machine learning. The mineral specimens used in this study were obtained from the Smithsonian Institution Mineralogy Collections Management Unit at the National Museum of Natural History (henceforth "NMNH"). The acquired mineral specimens are listed in **Table S1** and are grouped alphabetically by general mineral groupings. **Table S1** also lists the abbreviations used for each mineral as defined by @Whitney2010 unless otherwise specified. These minerals represent the majority of the most common silicate and aluminosilicate minerals that have been previously identified in ice core studies. Quartz and its polymorphs were not included because the simplicity of their EDS spectra makes them very easy to identify without the aid of a classification algorithm. 
+A total of 20 mineral standards were prepared in order to collect EDS data for training a novel mineral classification model using supervised machine learning. The mineral specimens used in this study were obtained from the Smithsonian Institution Mineralogy Collections Management Unit at the National Museum of Natural History (henceforth "NMNH"). The acquired mineral specimens are listed in **Table S1** and are grouped alphabetically by general mineral groupings. **Table S1** also lists the abbreviations used for each mineral as defined by Whitney & Evans (2010) unless otherwise specified. These minerals represent the majority of the most common silicate and aluminosilicate minerals that have been previously identified in ice core studies. Quartz and its polymorphs were not included because the simplicity of their EDS spectra makes them very easy to identify without the aid of a classification algorithm. 
 
 **Table S1**. Standard mineral specimens and their abbreviations. <sup>*a*</sup>Also belongs to the mica group. <sup>*b*</sup>Abbreviation defined for this study.
 
@@ -45,13 +45,13 @@ The Quattro SEM utilizes the EDAX APEX software for EDS analyses. This software 
 
 Because the minerals were powderized rather than polished, not every particle produced a clean EDS spectrum. This was by design, as opposed to polishing the samples, in order to train a machine learning algorithm capable of distinguishing mineral particles with non-ideal morphological features. However, some of the collected EDS data needed to be discarded due to poor signal detection that resulted from larger particles blocking x-rays from reaching the detector. Moreover, the chemical composition of the An mineral laregly overlapped with the other Ca-rich feldspars. Therefore, the An samples were discarded to avoid redundancy in the algorithm training. The Fs samples, likewise, were discarded but for a different reason. The elemental composition of the Fs samples did not match the expected compositon--either due to a misclassification of the hand sample or the presence of contaminants--and so the Fs data needed to be removed from algorithm training.
 
-The next step in preprocessing the mineral standard data was to calculate elemental ratios from the net intensity data measured on each mineral sample. Elemental ratios, as opposed to raw elemental net intensities, are strong predictors for discriminating between mineral species [@Donarummo2003]. A total of 23 elemental ratios were evaluated for each sample to be used as the predictor variables for training machine learning models (**Table S2**).
+The next step in preprocessing the mineral standard data was to calculate elemental ratios from the net intensity data measured on each mineral sample. Elemental ratios, as opposed to raw elemental net intensities, are strong predictors for discriminating between mineral species (Donarummo et al., 2003). A total of 23 elemental ratios were evaluated for each sample to be used as the predictor variables for training machine learning models (**Table S2**).
 
[…]

[…]
