### 2024-11-09 af492da
Truncations
