### 2024-11-09 #7422 @weber1158
@kthyng

- I added installation instructions to the `README.md` file. and a separate file under `MATLAB`>`docs`>`INSTALLATION.md`
- There is a test script/tutorial under `Tests`>`matlab_test.mlx`. An .mlx file is essentially MATLAB's version of a Jupyter Notebook. You can view the contents of the file online by navigating to the repository at [![View my project on File Exchange](https://www.mathworks.com/matlabcentral/images/matlab-file-exchange.svg)](https://www.mathworks.com/matlabcentral/fileexchange/170771) and clicking on the `Examples` tab.
- I have reduced the length of `paper.md` to 1027 words. However, I think this includes the words from the `Acknowledgements` and `Copyrights` sections. If so, then the body of the paper is only 871 words.
