### 2024-11-26 #7422 @rwegener2
👋🏻 @@dxm447 or @ThFriedrich, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
