### 2024-11-26 #7533 @rwegener2
👋🏼 @weber1158, @ThFriedrich, and @dxm447 this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```
