### 2024-12-02 4aad28d
Added a node visualization schematic to clarify local functions
