### 2024-12-06 3ff8cfa
Revised to more closely mirror the Matlab equivalent function
