+++ Julia/donarummo_classification.jl
@@ -15,26 +15,26 @@ INPUT
 
 OUTPUT
 
-"minerals" :: Categorical vector of mineral names corresponding to each row in the input DataFrame.
+"minerals" :: DataFrame of mineral names corresponding to each row in the input DataFrame.
 
 LIST OF POSSIBLE MINERAL CLASSIFICATIONS
 
- 1. Albite
- 2. Anorthite
- 3. Augite
- 4. Biotite
- 5. Chlorite
- 6. Hornblende
- 7. Hectorite
- 8. Illite
- 9. Illite/Smectite 70/30 mix
-10. Kaolinite
-11. Labradorite/Bytownite
-12. Montmorillonite
-13. Muscovite
-14. Oligoclase/Andesine
-15. Orthoclase
-16. Vermiculite
+ 1. Ab [Albite]
+ 2. An [Anorthite]
+ 3. Aug [Augite]
+ 4. Bt [Biotite]
+ 5. Chl [Chlorite]
+ 6. Hbl [Hornblende]
+ 7. Htr [Hectorite]
+ 8. Ilt [Illite]
+ 9. Ilt/Sme [Illite/Smectite 70/30 mix]
+10. Kln [Kaolinite]
+11. Lab/Byt [Labradorite/Bytownite]
+12. Mnt [Montmorillonite]
+13. Ms [Muscovite]
+14. Olig/Ans [Oligoclase/Andesine]
+15. Afs [Orthoclase = Alkali feldspar AKA Microcline, Mc]
+16. Vrm [Vermiculite]
 
 LIMITATIONS
 
@@ -77,438 +77,222 @@ function donarummo_classification(df)
             error("Input must be a DataFrame containing a column for Na, Mg, Al, Si, K, Ca, and Fe. Check the spellings of the column names. Only full element names and abbreviations are valid.")
         end
     
-        ###
-        ### CREATE LOCAL FUNCTIONS
-        ###
-        function mineral_classification(DF)
-            num_classifications = nrow(DF)
-            minerals = fill("Unknown", num_classifications) # Preallocate memory
-    
-            # FIRST BRANCHING PATHWAY
-                # HECTORITE
-                    idx = check_hectorite(DF)
-                    minerals[idx] .= ["Hectorite"]
-                # AUGITE
-                    idx = check_augite(DF)
-                    minerals[idx] .= ["Augite"]
-                # U-A
-                    idx = check_UA(DF)
-                    minerals[idx] .= ["U-A"]
-                # HORNBLENDE
-                    idx = check_hornblende(DF)
-                    minerals[idx] .= ["Hornblende"]
-            # SECOND BRANCHING PATHWAY - B
-                # ALBITE
-                    idx = check_albite(DF)
-                    minerals[idx] .= ["Albite"]
-                # OLIG/ANDESINE
-                    idx = check_olig(DF)
-                    minerals[idx] .= ["Olig/Andesine"]
-                # LAB/BYTOWNITE
-                    idx = check_lab(DF)
-                    minerals[idx] .= ["Lab/Bytownite"]
-                # U-B3
-                    idx = check_UB3(DF)
-                    minerals[idx] .= ["U-B3"]
-                # U-B2
-                    idx = check_UB2(DF)
-                    minerals[idx] .= ["U-B2"]
-                # Ca-MONTMORILLONITE
-                    idx = check_montmorillonite(DF)
-                    minerals[idx] .= ["Ca-Montmorillonite"]
[…]
