### 2024-12-06 d0f3835
Revised output description in function docs
