### 2024-12-11 9432f49
Improved data importing command
