### 2024-12-19 #7533 @ThFriedrich
The paper is well written and the repository clean, structured and well documented. I recommend accepting the article, though I may add a few minor comments:
### Formalities
 - Community Guidelines are missing
 - Some features don't seem to have a relevant context and I wonder why they were added, i.e.: [get_sem_metadata()](https://github.com/weber1158/eds-classification/blob/main/MATLAB/docs/DOCUMENTATION.md#get_sem_metadata), [sem_pixel_size()](https://github.com/weber1158/eds-classification/blob/main/MATLAB/docs/DOCUMENTATION.md#sem_pixel_size) and [convergence_angle()](https://github.com/weber1158/eds-classification/blob/main/MATLAB/docs/DOCUMENTATION.md#convergence_angle)
### Concept
 - The approach overall seems very solid
 - About the usage of SMOT for the generation of synthetic data: The empirical nature of SMOT seems fine in principle, but it generates **uncorrelated** random values in high-dimensional spaces, which could produce inconsistencies to real samples. The results suggest, that this is not an issue here, but perhaps one could check whether the validation accuracy drops if the validation set contains only real samples.
### Implementations
 - xray_peak_label function could be more sophisticated:
	 - multiple assigned labels per peak may be avoided by considering intensity ratios between e.g. Ka1 and Kb1 or ideally some kind of deconvolution, though this may be a bit much to ask for. I'd consider this optional, since it is clearly not a core functionality of the software
	 - tolerance in X is hardcoded and somewhat arbitrary. For example in the Albite example it causes to C-peak to not be identified

---

These are all quite minor suggestions. Overall I think this is very nice work, that is very accessible and easy to use. I'd be happy  accepting it as it is.
