### 2024-12-20 #7533 @weber1158
@ThFriedrich Thank you for taking the time to review my submission. I am glad you found that the functions in the repository are working properly, and I am especially happy that you find them easy to use. Making the functions as accessible as possible was a top priority of mine.

### Formalities
* I can add community guidelines to the documentation.
* I agree that the `get_sem_metadata`, `sem_pixel_size`, and `convergence_angle` functions aren't necessarily relevant to EDS research. I only included them because they are useful for other SEM applications. A few people in my department have been using them, so I decided that they were worth adding to the repository.

### Concept
* You make a good point about how SMOTE synthesizes uncorrelated data points. Luckily, it does not seem to affect the performance of the machine learning model: the accuracy when using only the measured data is 99.5%.

### Implementations
I really appreciate the feedback here. The `xray_peak_label` function has been a fun programming challenge for me, and I would love to improve the algorithm further.

* I have been trying to do exactly what you mentioned about using ratios to restrict the labeling to one element per peak, though this is still a work in progress. The original `xray_peak_label` function (in previous versions) only assigns one element to each peak, but this occasionally results in misclassifications because labels are preferentially assigned to the lowest atomic number element that emits a characteristic x-ray in the relevant energy range. My solution in the latest version was to label each peak with every possible element, which gives the user the ability to make an educated decision about which element is most likely. But like you say, it would be ideal if `xray_peak_label` was able to deduce the correct element automatically.

* I will definitely consider adding an optional input argument to set the tolerance in X. It should be easy to do. The only reason I did not do so before is because I wanted the algorithm to be as simplistic as possible for the user. Alternatively, I can see about making the function recursive so that if there are any major peaks that remain unlabeled the algorithm will run again using a higher tolerance.

---

Thank you again for your review and feedback!
