### 2024-12-29 63f46cc
Community guidelines initial commit
