+++ CONTRIBUTING.md
@@ -1,0 +1,39 @@
+# **Community Guidelines**
+
+### Contents
+* [Contribute to the software](#contribute)
+* [Report issues or problems with the software](#issues)
+* [Seek support](#support)
+
+## Contribute
+Contributions to the software are welcomed, including but not limited to:
+
+- Mineral classification algorithms transcribed from the peer-reviewed literature
+   - *Please include a citation to the original reference(s)*
+
+- Novel mineral classification algorithms/machine learning models that work with (1) EDS net intensity or atom percent data, (2) x-ray energy spectra, or (3) SEM micrographs.
+   - *Please include any associated model training data and scripts for reproducibility*
+
+- Homologous  functions designed for implementing the algorithms in this repository in other programming languages (e.g., `Python`, `Julia`, `R`)
+
+- Additional functions for working with x-ray energy spectra
+
+<!--
+### How to contribute:
+1. Fork the repository
+2. Create a new branch
+3. Make your changes
+4. Commit your changes
+5. Push to the branch
+6. Open a pull request
+-->
+
+## Issues
+### Fixes
+If you find a bug or documention errors, please open an issue [here](https://github.com/weber1158/eds-classification/issues). 
+
+### Suggestions
+Suggestions for improving the software, as well as feature requests, are welcomed [here](https://github.com/weber1158/eds-classification/issues).
+
+## Support
+Need help or have questions? Contact me by my institutional email `weber.1158@osu.edu`, or open a discussion [here](https://github.com/weber1158/eds-classification/discussions).
