### 2025-01-06 1debec0
Added optional input: `x_tolerance`
