### 2025-01-06 #7533 @weber1158
@rwegener2 Thank you for the reminder!

@ThFriedrich
- I have added a `CONTRIBUTING.md` file to the main branch that contains the community guidelines ([here](https://github.com/weber1158/eds-classification/blob/main/CONTRIBUTING.md)).
- I have also accepted your suggestion and added an optional input the the `xray_peak_label` function that allows the user to adjust the tolerance in x.
