### 2025-01-06 #7533 @dxm447
> @ThFriedrich thanks so much for your thoughtful feedback!
>
> @dxm447 do you have a timeline for when you anticipate being able to complete your review?
>
> @weber1158 Thanks for responding to the review comments! At JOSS the review process is iterative and the aim is to help you improve your software through community feedback. Because of this we ask that you take the time now to implement the feedback that the reviewers suggest. You are welcome to start an issue in your github repo if there is more detailed discussion to be had, just make sure to mention this review issue so I can track the progress.
>
> Please comment again when you have been able to address the changes that @ThFriedrich has proposed (ex. adding community guidelines, optional tolerance argument), so he can re-evaluate the JOSS review criteria. Thanks, and please reach out if you have any questions!

@rwegener2 I will be done by the middle of the week.
