### 2025-01-06 8409ea7
Updated `xray_peak_label` docs
