### 2025-01-14 1b6e57d
`xray_peak_label` and docs updates; and added private folder
