+++ MATLAB/private/xray_peak_label_best.m
@@ -1,0 +1,169 @@
+function xray_peak_label_best(axHandle,prominence,x_tolerance)
+%Label peaks in an xray_plot with the most likely ("best") element
+
+% Copyright ©Austin M. Weber 2025
+
+keV = axHandle.XData;
+Counts = axHandle.YData;
+min_sep = 0.1;
+
+% Delete any pre-existing labels
+points = findobj(gca,'Type','Line');
+labels = findobj(gca,'Type','Text');
+delete(points);
+delete(labels);
+
+% Check whether prominence was specified
+if ~exist('prominence','var') % i.e., if the user does not specify a prominence
+	% Evaluate the 90th-percentile cutoff point for determining local maxima
+	prominence = prctile(Counts,90);
+	% Find local maxima
+	[local_maxima_energies,idx] = evaluate_local_maxima(keV,Counts,min_sep,prominence);
+
+else % i.e., the user specifies a prominence
+	user_prominence = prctile(Counts,prominence);
+	[local_maxima_energies,idx] = evaluate_local_maxima(keV,Counts,min_sep,user_prominence);
+end
+
+% Check whether x_tolerance was specified
+if nargin == 3
+   % If true, is x_tolerance defined correctly?
+   if ~isa(x_tolerance,'double')
+	   error('The ''x_tolerance'' input must be numeric (default=0.025).')
+   end
+   if x_tolerance < 0
+	   error('The ''x_tolerance'' input must be greater than zero (default=0.025).')
+   end
+else
+   % Set x_tolerance
+   x_tolerance = 0.0250;
+end
+
+% Import x-ray energy table
+energies = readtable('XrayEnergyTable.csv','Delimiter','comma',...
+    'ReadVariableNames',true,'VariableNamingRule','preserve');
+element_names = energies.Element; % 95x1 cell
+element_Zs = energies.Z; % 95x1 double
+energy_vals = energies{:,3:end}; % 95x8 double
+
+% Loop through peaks
+peak_positions = find(idx);
+num_peaks = length(peak_positions);
+no_double_label=0; % Preallocate
+current_atomic_number = 1; % Preallocate
+for peak = 1:num_peaks
+ if no_double_label==peak
+	continue
+ end
+ nth_peak_energy = keV(peak_positions(peak));
+ % Find elements that may be associated with that peak
+ lowerBound = nth_peak_energy - x_tolerance;
+ upperBound = nth_peak_energy + x_tolerance;
+ % Match index matrix
+ match_matrix = energy_vals >= lowerBound & energy_vals <= upperBound;
+ % Get list of all matching elements (i.e., rows containing true values)
+ row_idx = logical(sum(match_matrix,2));
+ element_list = element_names(row_idx);
+ if peak == num_peaks
+	% Label the peak as the first element in the element list
+	hold('on')
+	% Add red marker to peak
+	plot_red_marker(local_maxima_energies,peak)
+	% Add label to peak
+	plot_element_label(local_maxima_energies,peak,element_list{1})
+	hold('off')
+ else
+	% Loop through the remaining peaks to check their element lists
+	for n=(peak+1):num_peaks
+ 	% Get the list of elements for the n+nth peak
+ 	n_plus_nth_peak_energy = keV(peak_positions(n));
+ 	lB = n_plus_nth_peak_energy - x_tolerance;
+ 	uB = n_plus_nth_peak_energy + x_tolerance;
+ 	mM = energy_vals >= lB & energy_vals <= uB;
+ 	rI = logical(sum(mM,2));
+ 	eL = element_names(rI);
+	% Delete the elements whose atomic numbers are less than the atomic
+	% number of the most recent element
+	atomic_numbers = element_Zs(rI);
[…]
