+++ MATLAB/xray_peak_label.m
@@ -1,148 +1,102 @@
 function xray_peak_label(axHandle,prominence,x_tolerance)
-%Label most prominent peaks in an xray_plot
+%Label peaks in an EDS spectrum
 %
 %Syntax
 % xray_peak_label(axHandle)
 % xray_peak_label(axHandle,prominence)
 % xray_peak_label(axHandle,prominence,x_tolerance)
+% xray_peak_label(axHandle,prominence,x_tolerance,algorithm)
+% xray_peak_label(axHandle,algorithm)
 %
 %Inputs
-% axHandle : axes handle to an open xray_plot object
-% prominence : (Optional, default=90) percentile between 0 and 100 that
-%  specifies the minimum prominence of the x-ray peaks to be labeled.
-% x_tolerance : (Optional, default=0.025) the energy value corresponding 
-%  to a peak in an EDS spectrum is unlikely to exactly match the literature 
-%  value for the element, so a tolerance in x must be defined. Tolerance 
-%  must be in units of keV. 
+% axHande (Required) : handle to an xray_plot axes
+% prominence (Optional) : minimum prominence of peaks to be labeled. Must
+%                         be a percentile between 1 and 99. Default=90
+% x_tolerance (Optional) : the +/- energy range used to identify elements
+%                          relative to the energy of each peak. Must be in
+%                          units of keV. Default=0.025
+% algorithm (Optional) : the labeling algorithm. Default='best'. The other
+%                        options are 'all' which labels each peak with all
+%                        of the possible elements, and 'first' which labels
+%                        each peak with the first possible element.
+%
 %
 %Examples
-% data = read_msa('file1.msa');
-% plt = xray_plot(data);
-% xray_peak_label(plt)
+% plt1 = xray_plot('file1.msa');
+% xray_peak_label(plt1)
 %
-% % Label the peaks above the 95th percentile in height
-% xray_peak_label(plt,95)
+% plt2 = xray_plot('file2.msa');
+% xray_peak_label(plt2,95,0.027,'all')
 %
 % See also
 %  xray_plot, read_msa
 
-% Copyright ©Austin M. Weber 2024
+% Copyright 2025 Austin M. Weber
 
-keV = axHandle.XData;
-Counts = axHandle.YData;
-min_sep = 0.1;
-
-% Delete any pre-existing labels
-points = findobj(gca,'Type','Line');
-labels = findobj(gca,'Type','Text');
-delete(points);
-delete(labels);
-
-% Check whether prominence was specified
-if ~exist('prominence','var') % i.e., if the user does not specify a prominence
-	% Evaluate the 90th-percentile cutoff point for determining local maxima
-	prominence = prctile(Counts,90);
-	% Find local maxima
-	[local_maxima_energies,idx] = evaluate_local_maxima(keV,Counts,min_sep,prominence);
-
-else % i.e., the user specifies a prominence
-	user_prominence = prctile(Counts,prominence);
-	[local_maxima_energies,idx] = evaluate_local_maxima(keV,Counts,min_sep,user_prominence);
+% Parse user inputs
+num_inputs = nargin;
+if num_inputs < 1
+	error('The function ''xray_peak_label'' requires at least one input.')
 end
 
-% Check whether x_tolerance was specified
-if nargin == 3
-   % If true, is x_tolerance defined correctly?
-   if ~isa(x_tolerance,'double')
-	   error('The ''x_tolerance'' input must be numeric (default=0.025).')
-   end
-   if x_tolerance < 0
-	   error('The ''x_tolerance'' input must be greater than zero (default=0.025).')
-   end
-else
[…]

[…]
