### 2025-01-14 d0fffcf
Add `algorithm` input to `xray_peak_label`
