### 2025-01-14 #7533 @weber1158
@ThFriedrich and @dxm447

I have updated the `xray_peak_label` function again with a new improvement: I added an optional input that allows the user to choose a labeling method. For example:

* `xray_peak_label(axHandle,"best")` will label each peak in an EDS spectrum using a best-guess method based on whether multiple K- and/or L-series lines for an element are present.

* `xray_peak_label(axHandle,"first")` will label each peak with the first possible element that the peak may represent according to atomic number.

* `xray_peak_label(axHandle,"all")` will label each peak with all of the possible elements that the peak could represent.

The function documentation has been updated accordingly [here](https://github.com/weber1158/eds-classification/blob/main/MATLAB/docs/DOCUMENTATION.md#xray_peak_label).
