### 2025-02-05 #7533 @kstenio
Hello @rwegener2, thanks for reaching out. It's been ages since the last time I worked with MATLAB, so I don't have a proper license right now. I'll check if my old institutional email will be eligible for a licence.

@weber1158 any chance this algorithm would work with GNU Octave?
