### 2025-02-05 #7533 @weber1158
@kstenio  Unfortunately, I do not think the software will be compatible with GNU Octave.

However, the software was designed in MATLAB Online, which is available to anyone for up to 20 hours a month when they create a [MathWorks account](https://www.mathworks.com/products/matlab-online/matlab-online-versions.html). No license required.
