### 2025-02-06 #7533 @kstenio
> [@kstenio](https://github.com/kstenio) Unfortunately, I do not think the software will be compatible with GNU Octave.
>
> However, the software was designed in MATLAB Online, which is available to anyone for up to 20 hours a month when they create a [MathWorks account](https://www.mathworks.com/products/matlab-online/matlab-online-versions.html). No license required.

I see. I did use my old institutional email for setting up an account, but seems like [UFSCar](https://en.wikipedia.org/wiki/Federal_University_of_S%C3%A3o_Carlos) does not have a global license agreement with MathWorks.

Either way, I'm somewhat available to do a review @rwegener2.
