### 2025-02-21 weber1158/eds-classification#1 @kstenio
JOSS Review - Issue #1 - Only a single author?
It is not impossible for a single person to be the author of a paper; however, it is not very common. Since most scientific development occurs within R&D organizations or universities, at least the name of a PI (Principal Investigator) is typically included. If anyone else contributed to this effort with you, even if they did not write code, I suggest adding their name as well.

Did you do all the work entirely alone?
