### 2025-02-21 weber1158/eds-classification#2 @kstenio
JOSS Review - Issue #2 - Links should use permalinks
Repositories are often transitional. For this reason, I recommend using permalinks when referencing them in the paper.
