### 2025-02-21 weber1158/eds-classification#3 @kstenio
JOSS Review - Issue #3 - Is the supplemental information an entirely new paper?
In the main paper, you state that the Weber algorithm is a novel method for EDS mineral classification, and then refer the reader to the supplemental material. There, your method is presented in what feels like an entirely separate paper, similar to a preprint. This additional paper is well written, and the results are promising, but I did not see any evidence of peer review.

I would recommend submitting this paper to another (non-software) journal, such as _Talanta_ or _Analytical Methods_. There is a good chance it would be accepted. Only then would I suggest using the term "novel."

_P.S.: This is more of a recommendation than an issue._
