### 2025-02-21 weber1158/eds-classification#4 @kstenio
JOSS Review - Issue #4 - Lack of instructions for running tests
I found the test folders, but I was unable to locate instructions on how to run them. The instructions and expected results should be clearer.
