### 2025-02-21 #7533 @kstenio
Hi @rwegener2 and @weber1158,

From my understanding, this paper meets many of the JOSS requirements, but a few areas could benefit from clarification or expansion. Specifically, there are points regarding authorship acknowledgment, references to previous work, and testing instructions that I believe should be addressed.

I’ve outlined my inquiries in the following issues:

* weber1158/eds-classification#1
* weber1158/eds-classification#2
* weber1158/eds-classification#3
* weber1158/eds-classification#4

That said, I would recommend this paper for publication.
