+++ Paper/paper.md
@@ -34,7 +34,7 @@ Therefore, to satisfy the needs of Earth and atmospheric scientists seeking to r
 # Methods
 ## Mineral Dust Classification
 
-The function library introduces a novel supervised machine learning model (the "Weber algorithm") that is trained to recognize 18 common minerals from EDS net intensity data and exhibits an overall accuracy exceeding 99% (see **[Supplemental Information](https://github.com/weber1158/eds-classification-for-matlab/blob/main/Paper/supplement.md)**). The library also includes functions for three non-machine learning algorithms that have been painstakingly transcribed from the literature into MATLAB and Julia code. This includes the sorting algorithm from @Donarummo2003 which can be used to classify 16 unique aluminosilicate minerals based on EDS net intensity data, the Panta algorithm [@Panta2023] which can be used to classify 17 unique minerals and 6 mineral groups based on EDS atom percent data, and the Kandler algorithm [@Kandler2011] which can be used to categorize EDS atom percent data into broad mineral groups, classes, and refractive indexes. 
+The function library introduces a supervised machine learning model (the "Weber algorithm") that is trained to recognize 18 common minerals from EDS net intensity data and exhibits an overall accuracy exceeding 99% (see **[Supplemental Information](https://github.com/weber1158/eds-classification/blob/902b82aa5b984983c448366cbf4794b62273d765/Paper/supplement.md)**). The library also includes functions for three non-machine learning algorithms that have been painstakingly transcribed from the literature into MATLAB and Julia code. This includes the sorting algorithm from @Donarummo2003 which can be used to classify 16 unique aluminosilicate minerals based on EDS net intensity data, the Panta algorithm [@Panta2023] which can be used to classify 17 unique minerals and 6 mineral groups based on EDS atom percent data, and the Kandler algorithm [@Kandler2011] which can be used to categorize EDS atom percent data into broad mineral groups, classes, and refractive indexes. 
 
 The four mineral classification algorithms are available as individual functions (`weber_classification`, `donarummo_classifiation`,  `panta_classifiation`, and `kandler_classification`) as well as packaged together in a single function (`eds_classification`). Each function has clearly written documentation that can be viewed using the MATLAB `help` or `doc` commands. The general syntax, however, is very basic:
 
