### 2025-02-21 weber1158/eds-classification#1 @weber1158
@kstenio  Indeed, thus far, I am the only contributor to the repository in terms of: Conceptualization, Data Curation, Analysis, Investigation, Software, Validation, and Writing. My PhD advisor has supported the project, but is not an active contributor with respect to the development of the software, the repository, or writing.

The Acknowledgements section in the paper does give thanks to my advisor and to a few other colleagues who either inspired me or helped teach me the concepts that went into developing the software. I have also just (slightly) expanded the Acknowledgements section to include a few more individuals that I would like to recognize; but again, no one else has formally participated in the creation of the software or the repository.

The original purpose of the software was to assist in the analysis of ice core data that I will be collecting for my dissertation research. However, because its potential applications extend beyond my own scientific endeavors, I wanted to make the software available to others as an open-source resource.
