### 2025-02-21 weber1158/eds-classification#2 @weber1158
@kstenio  I agree that using permalinks would be wise.

It turns out that this was very easy to do on GitHub—all I needed to do was type `y` with the file open and the URL box updated to a permalink address for the specific commit that I was viewing. I simply copied this new link and replaced the old link in the `paper.md` file.
