### 2025-02-21 weber1158/eds-classification#3 @weber1158
@kstenio  The supplemental information is not meant to be a separate paper; rather, I truly meant it as supplementary material. I was treating the supplement more like documentation to help explain how the machine learning model was developed.

I was expecting that any peer-review would come through my JOSS submission since all of the data and functions needed to replicate the results are available in the repository. There is even a step-by-step walkthrough [here]( https://github.com/weber1158/eds-classification/blob/d0fffcf8afc68393e10fef563c33f2e4fa005b75/MATLAB/MachineLearningModel/weber_algorithm_training.m).

But if you think the material in the supplemental information warrants its own, more technical review (and also fits the scope of *Talanta*) then I will consider writing a more formal manuscript to submit there in addition to the current JOSS submission. I appreciate the recommendation!

As for how the machine learning model is described in the main paper, I have removed the word “novel” per your suggestion.
