### 2025-02-21 weber1158/eds-classification#4 @weber1158
@kstenio  The test file for the MATLAB software is a Live Script (`.mlx`), which is MATLAB’s version of a Jupyter Notebook. Live scripts open just like any other MATLAB (`.m`) file, and each section in the live script can be run individually (<kbd>Ctrl</kbd>+<kbd>Enter</kbd>) or all at once.

You can also view the output of the test file without evening opening it.  On the MATLAB Central File Exchange you can simply navigate to the software repository [here]( https://www.mathworks.com/matlabcentral/fileexchange/170771) and click on the Examples tab. It’s quite convenient, but you’re right that it may not be as self-explanatory as I had hoped. I have added a statement to the `README` that directs the user on how to view the tests.

In addition, I have migrated the test file into the MATLAB folder because I think it makes more sense for its location to be in the same place as the functions that it calls.
