### 2025-02-21 #7533 @weber1158
@kstenio I appreciate you taking the time to review my software submission.

You will find my responses to your feedback on the main repository's [Issues page](https://github.com/weber1158/eds-classification/issues).  My revisions to `paper.md` seem to be compiling correctly (see my `@ editorialbot generate pdf` command above or click [here](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07533/joss.07533/10.21105.joss.07533.pdf) to download a PDF copy).

Please let me know if I have addressed each issue to your satisfaction so that I may mark the issues as "closed". Thank you again for your time!
