### 2025-02-22 weber1158/eds-classification#1 @kstenio
Great. For me your motivation is sufficient. :-)
I'll close the issue
