### 2025-02-22 weber1158/eds-classification#3 @kstenio
There's no problem in submitting as is (the supplemental material, I mean), but it has a nice approach for the problem you want to solve. And, as an additional reference for any other journal, you may reference the Joss paper, as it will be published soon :-)
