### 2025-03-05 #7533 @rwegener2
Hi @weber1158, thanks for a nice paper! This is a great tool and the community is lucky to have it. A few comments on the paper:

* The Summary of a JOSS paper should read a bit more like an academic abstract. It should have a sentence or two introducing the role of SEM EDS, a sentence or two about the gap that exists in the tooling, then a few sentences about the way that this tool fills the gap. I think you already address well what this tool does, but the introduction and description of gap should be added
* line 20: Please define both SEM and EDS in a bit more detail. Re-stating the acronym would be good, and EDS could use a sentence briefly describing the technique
* line 31: I suggest you put Earth in lowercase
* line 33: What is a "scripting algorithm"? Does this refer to a specific type of algorithm? Or does it simply refer to set of scripts that implement different EDS algorithms? The phrasing makes it sound like the first one (a specific type of algorithm), so I would either define "scripting algorithm" if that is what it is, or rephrase if it is the second one (a set of scripts implementing algorithms)
* your methods descriptions are really well done!
