### 2025-03-05 8c81885
Removed the word 'novel' per recommendation by reviewer
