### 2025-03-05 c0f907e
Abstract revisions
