+++ Paper/paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-A repository of MATLAB functions has been assembled to work with scanning electron microcope (SEM) and energy dispersive x-ray spectrometry (EDS) data, with the objective of providing researchers with tools to quickly and accurately determine the mineral compositions of dust particles. The repository includes a novel machine learning classifier and three additional sorting algorithms that have been transcribed from the peer-reviewed literature, as well as functions for importing and visualizing x-ray energy spectra.
+Mineral dust plays an important role in governing Earth's energy balance and also accelerates the melting of glaciers. The most common analytical method for studying mineral dust is scanning electron microscopy (SEM) with energy dispersive x-ray spectrometry (EDS). The primary benefit of SEM-EDS over alternative techniques is that it allows researchers to select and measure the elemental compositions of individual particles. However, elemental composition alone is often not enough to confidently infer mineralogy because many unrelated minerals are comprised of the same elements. Identifying minerals from SEM-EDS data has also historically been a labor intensive task because automated procedures are not readily available. To address these limitations, a repository of MATLAB functions has been assembled to work with SEM-EDS data, with the objective of providing researchers with tools to quickly and accurately determine the mineral compositions of dust particles. The repository includes a machine learning classifier and three additional sorting algorithms that have been transcribed from the peer-reviewed literature, as well as functions for importing and visualizing x-ray energy spectra.
 
 # Statement of need
 The mineral classification of single particles (i.e., dust) has major implications for atmospheric and glaciological research. For instance, there is a great need for building mineral dust datasets that can be used to improve the predictive ability of climate models [@Elements2010]. Mineral dust observations are also severely limited in many glaciated mountain regions, meaning that the dust's radiative impact on glacier melt rates cannot be easily determined [@Gilardoni2022].
