+++ Paper/paper.md
@@ -27,9 +27,9 @@ Mineral dust plays an important role in governing Earth's energy balance and als
 # Statement of need
 The mineral classification of single particles (i.e., dust) has major implications for atmospheric and glaciological research. For instance, there is a great need for building mineral dust datasets that can be used to improve the predictive ability of climate models [@Elements2010]. Mineral dust observations are also severely limited in many glaciated mountain regions, meaning that the dust's radiative impact on glacier melt rates cannot be easily determined [@Gilardoni2022].
 
-One of the most common methods for reconstructing the mineral dust history of a location or region has been the use of SEM-EDS to analyze single particles extracted from ice cores [@Donarummo2003; @Wu2016; @Nagatsuka2021]. However, the process of classifying individual particles based on their EDS data is highly laborious and prone to error due to the complexity of certain minerals' elemental signatures. The adoption of numerous classification techniques is typically required to improve the certainty of the classifications before finalizing the results. Many have used the dichotomous key in @Severin2004 to compare EDS spectra to that of mineral standards, but this protocol requires access to a physical or digital copy of the book and cannot be automated. Others have used published sorting algorithms like the one by @Donarummo2003, but these algorithms are limited in the number of minerals that they can identify and, to the author's knowledge, each of these algorithms is only available in print and not as a functional computer program. 
+One of the most common methods for reconstructing the mineral dust history of a location or region has been the use of scanning electron microscopy with energy dispersive x-ray spectrometry (SEM-EDS) to analyze single particles extracted from ice cores [@Donarummo2003; @Wu2016; @Nagatsuka2021]. SEM allows researchers to examine particles individually while an EDS detector measures the particles' elemental compositions as they emit x-rays under the influence of an electron beam. However, the process of classifying single particles from EDS data is highly laborious and prone to error due to the complexity of certain minerals' elemental signatures. The adoption of numerous classification techniques is typically required to improve the certainty of the classifications before finalizing the results. Many have used the dichotomous key in @Severin2004 to compare EDS spectra to that of mineral standards, but this protocol requires access to a physical or digital copy of the book and cannot be automated. Others have used published sorting algorithms like the one by @Donarummo2003, but these algorithms are limited in the number of minerals that they can identify and, to the author's knowledge, each of these algorithms is only available in print and not as a functional computer program. 
 
-Therefore, to satisfy the needs of Earth and atmospheric scientists seeking to rapidly classify mineral particles from EDS data with a high degree of certainty, an open-source repository of relevant scripting algorithms has been developed for MATLAB (along with several homologous functions written in the Julia programming language). This library of functions gives users the tools to work with SEM-EDS data and to perform mineral classification with ease.
+Therefore, to satisfy the needs of Earth and atmospheric scientists seeking to rapidly classify mineral particles from EDS data with a high degree of certainty, an open-source repository of functions for implementing different mineral classification algorithms has been developed for MATLAB (along with several homologous functions written in the Julia programming language). This library of functions gives users the tools to work with SEM-EDS data and to perform mineral classification with ease.
 
 # Methods
 ## Mineral Dust Classification
