### 2025-03-05 #7533 @weber1158
Hi @rwegener2 ,

I’m happy you think the paper will be a great tool for the community!

My apologies for not organizing the Summary more like an abstract. I was trying to keep it short and concise to satisfy the journal’s word count limit. The Summary is now updated and includes: (i) a sentence introducing the significance of mineral dust; (ii) two sentences introducing SEM-EDS; and (iii) two sentences explaining the gaps/limitations of SEM-EDS that the software seeks to address.

Line 20: I have defined SEM-EDS and provided a terse description of its use.

Line 31:  I agree that “Earth and atmospheric” looks a little odd because only one word is capitalized. Although, in this context I do think Earth needs to be capitalized because it is referring to the field of Earth science, which is usually written with an uppercase E. Under the circumstances, do you think this is acceptable?

Line 33: Now that I’m looking at it, I agree that “scripting algorithm” sounds ambiguous and does not convey what I was trying to say. I have rephrased the sentence as follows: “…an open-source repository of ~~relevant scripting algorithms~~ functions for implementing different mineral classification algorithms has been developed for MATLAB …”.
