### 2025-03-07 #7533 @rwegener2
Looks great @weber1158! Thanks for the updates.

At this point please go ahead and complete the tasks in section 2 of [this comment above](https://github.com/openjournals/joss-reviews/issues/7533#issuecomment-2695625215). You can post the DOI and version in this issue.
