### 2025-03-07 #7533 @rwegener2
Alright @weber1158  I have completed my checks and recommended this paper for publication. I appreciate your dedication to getting feedback and improving your software, it's a great contribution to the community. Thank you to @kstenio and @ThFriedrich for your thoughtful reviews!

At this point the Track Editor will perform a final set of checks. This is the last step in the publication process. As always feel free to reach out in this Github issue if you have any questions.
