### 2025-03-10 #7533 @weber1158
@rwegener2, @ThFriedrich, @kstenio, and @kthyng:

Wow, thank you all for the pleasant review experience! I really appreciate the time and consideration you have given me. I'm looking forward to adding my name to the list of potential reviewers for future *JOSS* submissions!
