### 2025-03-10 #7533 @kstenio
> [@rwegener2](https://github.com/rwegener2), [@ThFriedrich](https://github.com/ThFriedrich), [@kstenio](https://github.com/kstenio), and [@kthyng](https://github.com/kthyng):
>
> Wow, thank you all for the pleasant review experience! I really appreciate the time and consideration you have given me. I'm looking forward to adding my name to the list of potential reviewers for future _JOSS_ submissions!

Congratulations for the paper @weber1158! 🥳
